Refacto Update and add delete Quiz
