+++ src/components/Teacher/TeacherQuiz.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { useQuery, gql } from '@apollo/client'
+import { useQuery, gql, useMutation } from '@apollo/client'
 import { Link } from 'react-router-dom'
 
 // MUI Import
@@ -27,6 +27,11 @@ const ALL_QUIZZES = gql`
     }
   }
 `
+const DELETE_QUIZ = gql`
+  mutation deleteQuiz($id: String) {
+    deleteQuiz(id: $id)
+  }
+`
 
 const useStyles = makeStyles(() =>
   createStyles({
@@ -47,22 +52,36 @@ const TeacherQuiz = (): JSX.Element => {
     data: dataQuizzes
   } = useQuery(ALL_QUIZZES)
 
+  const [deleteQuiz] = useMutation(DELETE_QUIZ, { refetchQueries: ['quizzes'] })
+
+  const removeQuiz = async (idQuiz: string) => {
+    try {
+      await deleteQuiz({
+        variables: {
+          id: idQuiz
+        }
+      })
+    } catch (e) {
+      console.error(e)
+    }
+  }
+
   if (loadingQuizzes) {
     return <p>Loading...</p>
   }
   if (errorQuizzes) {
     return <p>Error...</p>
   }
-  // const quizzesDataWritable = JSON.parse(JSON.stringify(dataQuizzes))
+  const quizzesDataWritable = JSON.parse(JSON.stringify(dataQuizzes))
 
   return (
     <div className="teacherHome-container">
       <p>Je suis la page quiz du Teacher</p>
-      <Button component={Link} to="/quiz/create">
+      <Button component={Link} to="/quiz/editor">
         Créer un quiz
       </Button>
-      {dataQuizzes.quizzes &&
-        dataQuizzes.quizzes.map((quiz: IQuiz) => (
+      {quizzesDataWritable.quizzes &&
+        quizzesDataWritable.quizzes.map((quiz: IQuiz) => (
           <div className={classes.quizItem}>
             <p key={quiz.title}>{quiz.title}</p>
             <div>
@@ -78,7 +97,7 @@ const TeacherQuiz = (): JSX.Element => {
                   <EditOutlinedIcon />
                 </IconButton>
               </Link>
-              <IconButton>
+              <IconButton onClick={() => removeQuiz(quiz.id)}>
                 <DeleteOutlineIcon />
               </IconButton>
             </div>
