+++ src/components/Teacher/TeacherQuizEditor.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-dom'
 import { useMutation, gql } from '@apollo/client'
 
 // MUI Import
-import { makeStyles, createStyles, Theme } from '@material-ui/core/styles'
+import { makeStyles, createStyles } from '@material-ui/core/styles'
 import {
   Paper,
   Input,
@@ -17,7 +17,23 @@ import CloseIcon from '@material-ui/icons/Close'
 import DeleteOutlineIcon from '@material-ui/icons/DeleteOutline'
 
 // Type
-import { IQuestion, IQuiz, IResponse } from '../../types/quiz'
+import { IQuestion, IResponse } from '../../types/quiz'
+
+const ADD_QUIZ = gql`
+  mutation createQuiz($quiz: inputQuiz) {
+    createQuiz(quiz: $quiz) {
+      title
+      comment
+      questions {
+        question
+        responses {
+          response
+          isCorrect
+        }
+      }
+    }
+  }
+`
 
 const UPDATE_QUIZ = gql`
   mutation updateQuiz($id: String, $quiz: inputQuiz) {
@@ -36,7 +52,7 @@ const UPDATE_QUIZ = gql`
   }
 `
 
-const useStyles = makeStyles((theme: Theme) =>
+const useStyles = makeStyles(() =>
   createStyles({
     quiz: {
       display: 'flex',
@@ -82,23 +98,41 @@ const useStyles = makeStyles((theme: Theme) =>
 const TeacherQuizEditor = (props: any): JSX.Element => {
   const classes = useStyles()
   const history = useHistory()
+  const [addQuiz] = useMutation(ADD_QUIZ, { refetchQueries: ['quizzes'] })
   const [updateQuiz] = useMutation(UPDATE_QUIZ, { refetchQueries: ['quizzes'] })
   const [idQuiz, setIdQuiz] = useState()
-  const [title, setTitle] = useState()
+  const [title, setTitle] = useState<string>('')
   const [comment, setComment] = useState()
   const [questions, setQuestions] = useState<IQuestion[]>([])
 
   useEffect(() => {
-    const propsQuiz = props.location.state.quiz
-    console.log(propsQuiz)
-    setIdQuiz(propsQuiz.id)
-    setTitle(propsQuiz.title)
-    if (propsQuiz.comment) {
-      setComment(propsQuiz.comment)
+    if (props.location.state !== undefined) {
+      const propsQuiz = props.location.state.quiz
+      setIdQuiz(propsQuiz.id)
+      setTitle(propsQuiz.title)
+      if (propsQuiz.comment) {
+        setComment(propsQuiz.comment)
+      }
+      setQuestions(propsQuiz.questions)
     }
-    setQuestions(propsQuiz.questions)
     // eslint-disable-next-line
   }, [])
+
+  const postQuiz = async () => {
+    try {
+      await addQuiz({
+        variables: {
+          quiz: {
+            title,
+            comment,
+            questions
+          }
+        }
+      })
+    } catch (e) {
+      console.error(e)
+    }
+  }
 
   const editQuiz = async () => {
     try {
@@ -266,13 +300,16 @@ const TeacherQuizEditor = (props: any): JSX.Element => {
         onClick={async (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
           e.preventDefault()
           try {
-            editQuiz()
+            if (props.location.state !== undefined) {
+              editQuiz()
+            } else {
+              postQuiz()
+            }
             history.goBack()
           } catch (error) {
             console.error(error)
           }
         }}
-        // onClick={() => console.log(questions)}
       >
         Enregistrer le quiz
       </Button>
